Upgrade to non-deprecated protobuf library
github.com/golang/protobuf is deprecated, and has been replaced by google.golang.org/protobuf. Migrating to this library required recompiling the proto file - I also updated the package name to be unique as advised in the FAQ (https://protobuf.dev/reference/go/faq/#namespace-conflict), and added a Makefile task to record how to re-compile.

I've confirmed that the wire format of the messages is the same before and after migrating and re-compilation, so clients using older versions of the protobuf library will still be able to unmarshal payloads generated using this version.